# result location mechanism (part of no-copy semantics)