--- a/BRANCH_TODO
+++ b/BRANCH_TODO
@@ -1,6 +1,8 @@
 Scratch pad for stuff to do before merging master
 =================================================
 
+uncomment all the behavior tests
+
 restore test_runner.zig to master branch
  - also the default panic function and unexpected_error_tracing. see the commit
    that adds this text to BRANCH_TODO file.
@@ -8,11 +10,21 @@ restore test_runner.zig to master branch
 
 get an empty file compiling successfully (with no panic fn override)
 
-uncomment all the behavior tests
-
 better behavior for implicit casts. for example these introduce an extra allocation/memcpy:
     var x: [1]i32 = [_]i32{1};
     var x = ([1]i32)([_]i32{1});
 whereas this one does not:
     var x = [_]i32{1};
 but all 3 should be semantically identical
+
+
+This example has less than ideal LLVM IR:
+```zig
+export fn entry() void {
+    _ = mul(true, 1) catch undefined;
+}
+pub fn mul(c: bool, answer: i32) error{Overflow}!i32 {
+    return if (c) error.Overflow else answer;
+}
+```
+It creates an unnecessary stack variable.