--- a/BRANCH_TODO
+++ b/BRANCH_TODO
@@ -2,6 +2,7 @@ Scratch pad for stuff to do before merging master
 =================================================
 
 uncomment all the behavior tests
+diff master branch to make sure
 
 restore test_runner.zig to master branch
  - also the default panic function and unexpected_error_tracing. see the commit