--- a/BRANCH_TODO
+++ b/BRANCH_TODO
@@ -4,10 +4,7 @@ Scratch pad for stuff to do before merging master
 uncomment all the behavior tests
 diff master branch to make sure
 
-restore test_runner.zig to master branch
- - also the default panic function and unexpected_error_tracing. see the commit
-   that adds this text to BRANCH_TODO file.
- - and std/specia/bootstrap.zig
+restore bootstrap.zig to master
 
 get an empty file compiling successfully (with no panic fn override)