--- a/src/all_types.hpp
+++ b/src/all_types.hpp
@@ -3099,6 +3099,7 @@ struct IrInstructionTestErr {
 struct IrInstructionUnwrapErrCode {
     IrInstruction base;
 
+    bool initializing;
     IrInstruction *err_union_ptr;
 };