--- a/src/all_types.hpp
+++ b/src/all_types.hpp
@@ -335,8 +335,9 @@ struct ConstExprValue {
         RuntimeHintSlice rh_slice;
     } data;
 
-    ConstExprValue(const ConstExprValue &other) = delete; // plz zero initialize with {}
-    ConstExprValue& operator= (const ConstExprValue &other) = delete; // use copy_const_val
+    // uncomment these to find bugs. can't leave them uncommented because of a gcc-9 warning
+    //ConstExprValue(const ConstExprValue &other) = delete; // plz zero initialize with {}
+    //ConstExprValue& operator= (const ConstExprValue &other) = delete; // use copy_const_val
 };
 
 enum ReturnKnowledge {