… src/ir.cpp
… src/ir_print.cpp
… std/event/fs.zig
… std/event/lock.zig
… std/event/net.zig
… std/event/rwlock.zig
… std/heap.zig
… std/io/test.zig
… std/json.zig
… std/os.zig
… std/os/windows.zig
… std/special/bootstrap.zig
… std/special/panic.zig
… std/special/test_runner.zig
… std/std.zig
… std/zig/parser_test.zig
… std/zig/render.zig
… test/compile_errors.zig
… test/stage1/behavior.zig
… test/stage1/behavior/bitcast.zig
… test/stage1/behavior/cast.zig
… test/stage1/behavior/enum.zig
… test/stage1/behavior/error.zig
… test/stage1/behavior/eval.zig
… test/stage1/behavior/fn.zig
… test/stage1/behavior/for.zig
… test/stage1/behavior/generics.zig
… test/stage1/behavior/if.zig
… test/stage1/behavior/misc.zig
… test/stage1/behavior/optional.zig