* Previous effort (1) #1652
 * Previous effort (2) #1682

This is the third attempt, which is ~working beautifully~ working.

After this is done, the door is opened to some important core language improvements, most notably the coroutine rewrite (#2377) and well-defined copy eliding semantics (#287). The main idea here is that you can look at any expression and easily understand where the result will be written to. For example:

```zig
var x = foo();
```

With the new result location semantics, the function call `foo()` is guaranteed to directly initialize the memory that the variable `x` is associated with. Note that this is a semantic guarantee, regardless of optimization level. What's neat about this is that it works with complex expressions, for example:

```zig
var x = Data{
    .field1 = 26,
    .field2 = if (condition) bar() else baz(),
};
```

Here, the function calls to `bar()` and `baz()` are guaranteed to directly initialize `x.field2`. Have a look at the earlier commit messages for more demonstrations of zig code and the corresponding LLVM IR it's generating.

Merge checklist:
 * [x] Get compiler_rt tests passing
 * [x] Get the behavioral tests passing
 * [x] Fix labeled break from a block to use proper result location semantics
 * [x] Get libuserland.a building successfully
 * [x] Fix stack trace / debug info regression
 * [x] Get standard library tests passing. Do a git diff with master to make sure all tests are uncommented.
 * [x] Fix the vector alignment issue that @mikdusan [found](https://gist.github.com/mikdusan/6e0d22f18f269ad87d2b2fa65db3a545#file-copy-elision-3-ll-L33)
 * [x] Fix the union initialization issue that @mikdusan [found](https://gist.github.com/mikdusan/1961bb9258740d0bf7a11e01e7d5c0bf)
 * [x] Fix compile errors with self hosted compiler ([@mikdusan's reduction](https://gist.github.com/mikdusan/f99c91734cc6ac0da7a07ccf47469d4d))
 * [ ] ~Get the self-hosted compiler building~ I'm allowing this to regress until #2377 is done.
 * [x] Merge master & resolve conflicts
 * [x] Get compile error tests passing
 * [x] All tests passing on Linux
 * [x] All tests passing on macOS
 * [x] All tests passing on Windows
 * [x] Work with @shawnl to get #2709 merged into master so that I can be the one to resolve the conflicts
 * [x] Offer to run the test suite for some zig community members' projects to find more regressions

Some more items, which I may postpone by filing issues rather than doing them in this branch:

 * [ ] better behavior for implicit casts. for example these introduce an extra allocation/memcpy:
```zig
    var x: [1]i32 = [_]i32{1};
    var x = ([1]i32)([_]i32{1});
```
whereas this one does not:
```zig
    var x = [_]i32{1};
```
but all 3 should be semantically identical.

 * [ ] This example has less than ideal LLVM IR:
```zig
export fn entry() void {
    _ = mul(true, 1) catch undefined;
}
pub fn mul(c: bool, answer: i32) error{Overflow}!i32 {
    return if (c) error.Overflow else answer;
}
```
It creates an unnecessary stack variable.

#### Preliminary Performance Comparison

```
result location branch, debug zig build, 622 behavioral tests

                Name       Start         End    Duration     Percent
          Initialize      0.0000      0.0035      0.0035      0.0009
   Semantic Analysis      0.0035      1.1861      1.1826      0.3220
     Code Generation      1.1861      1.3654      0.1793      0.0488
    LLVM Emit Output      1.3654      3.6289      2.2635      0.6163
  Build Dependencies      3.6289      3.6302      0.0013      0.0003
           LLVM Link      3.6302      3.6728      0.0426      0.0116
               Total      0.0000      3.6728      3.6728      1.0000

master branch, debug zig build, 616 behavioral tests

                Name       Start         End    Duration     Percent
          Initialize      0.0000      0.0026      0.0026      0.0008
   Semantic Analysis      0.0026      0.8965      0.8939      0.2586
     Code Generation      0.8965      1.0744      0.1779      0.0514
    LLVM Emit Output      1.0744      3.4139      2.3395      0.6767
  Build Dependencies      3.4139      3.4158      0.0019      0.0006
           LLVM Link      3.4158      3.4571      0.0413      0.0120
               Total      0.0000      3.4571      3.4571      1.0000
```

It's not a fair comparison because more tests have been added in the result location branch. However, note that the percent of time spent in LLVM Emit Output is a smaller percent, and the time spent in Semantic Analysis is a higher percent. This is a good thing, because it means that Zig's IR code - which I hastily coded up without regard for performance in this branch - can be more easily improved than LLVM's perf. The LLVM IR that this branch produces is significantly simpler for LLVM to compile, and significantly fewer calls to memcpy at runtime.