@mikdusan 2019-06-24
#### patch: vector alignment issue

add alignment to (vector → array) store result location

```diff
diff --git a/src/codegen.cpp b/src/codegen.cpp
index 4eea729d..e12f8975 100644
--- a/src/codegen.cpp
+++ b/src/codegen.cpp
@@ -5451,7 +5451,7 @@ static LLVMValueRef ir_render_vector_to_array(CodeGen *g, IrExecutable *executab
     LLVMValueRef vector = ir_llvm_value(g, instruction->vector);
     LLVMValueRef casted_ptr = LLVMBuildBitCast(g->builder, result_loc,
             LLVMPointerType(get_llvm_type(g, instruction->vector->value.type), 0), "");
-    gen_store_untyped(g, vector, casted_ptr, 0, false);
+    gen_store_untyped(g, vector, casted_ptr, get_ptr_align(g, instruction->result_loc->value.type), false);
     return result_loc;
 }
 ```

#### reduction: foo.zig

```zig
export fn foobar() void {
    var a: @Vector(4, i32) = [_]i32{ 1, 2, 3, 4 };
    var result_array: [4]i32 = a;
}
```

#### patched llvm-ir

```llvm
define void @foobar() #2 !dbg !35 {
Entry:
  %a = alloca <4 x i32>, align 16
  %result_array = alloca [4 x i32], align 4
  store <4 x i32> <i32 1, i32 2, i32 3, i32 4>, <4 x i32>* %a, align 16, !dbg !47
  call void @llvm.dbg.declare(metadata <4 x i32>* %a, metadata !39, metadata !DIExpression()), !dbg !47
  %0 = load <4 x i32>, <4 x i32>* %a, align 16, !dbg !48
  %1 = bitcast [4 x i32]* %result_array to <4 x i32>*, !dbg !48
  store <4 x i32> %0, <4 x i32>* %1, align 4, !dbg !48
  call void @llvm.dbg.declare(metadata [4 x i32]* %result_array, metadata !45, metadata !DIExpression()), !dbg !49
  ret void, !dbg !50
}
```

#### original llvm-ir

```llvm
define void @foobar() #2 !dbg !35 {
Entry:
  %a = alloca <4 x i32>, align 16
  %result_array = alloca [4 x i32], align 4
  store <4 x i32> <i32 1, i32 2, i32 3, i32 4>, <4 x i32>* %a, align 16, !dbg !47
  call void @llvm.dbg.declare(metadata <4 x i32>* %a, metadata !39, metadata !DIExpression()), !dbg !47
  %0 = load <4 x i32>, <4 x i32>* %a, align 16, !dbg !48
  %1 = bitcast [4 x i32]* %result_array to <4 x i32>*, !dbg !48
  store <4 x i32> %0, <4 x i32>* %1, !dbg !48
  call void @llvm.dbg.declare(metadata [4 x i32]* %result_array, metadata !45, metadata !DIExpression()), !dbg !49
  ret void, !dbg !50
}
```