@andrewrk 2019-06-24
Brilliant, thanks @mikdusan for the patch, looks good. Do you want to make a pull request so you can get credit for it (you can select copy-elision-3 instead of master as your destination branch)?