@mikdusan 2019-06-24
While making patch 3021e5ca67acca6cf7420bc5e2400aa6965596f9 I came across a compiler segfault which is a copy-elision-3 regression involving vector copy to array.

Please look over (patch is in comments) https://gist.github.com/mikdusan/642e780f6fc63c29dfbda1e091d63fd3