@shawnl 2019-06-25
../src/analyze.cpp:4716:63: error: ‘void* memcpy(void*, const void*, size_t)’ writing to an object of type ‘struct ConstExprValue’ with no trivial copy-assignment [-Werror=class-memaccess]
         memcpy(const_val, entry->value, sizeof(ConstExprValue));